lessons 108 - 109: determining available tables
