+++ pages/api/restaurant/[slug]/availability.ts
@@ -2,12 +2,7 @@ import { PrismaClient } from "@prisma/client";
 import { table } from "console";
 import { NextApiRequest, NextApiResponse } from "next";
 import { times } from "../../../../data";
-
-type BookingTablesObjectType = {
-  [key: string]: {
-    [key: number]: true;
-  };
-};
+import { findAvailableTables } from "../../../../services/restaurant/findAvailableTables";
 
 const prisma = new PrismaClient();
 
@@ -23,39 +18,6 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     res.status(400).json({ message: "Invalid data provided" });
     return;
   }
-
-  const searchTimes = times.find((t) => t.time === time)?.searchTimes; //JOKE: t.time = 17:00 o'clock xD!
-
-  if (!searchTimes) {
-    res.status(400).json({ message: "Invalid time provided" });
-    return;
-  }
-
-  const bookings = await prisma.booking.findMany({
-    where: {
-      booking_time: {
-        gte: new Date(`${day}T${searchTimes[0]}`),
-        lte: new Date(`${day}T${searchTimes[searchTimes.length - 1]}`),
-      },
-    },
-    select: {
-      booking_time: true,
-      number_of_people: true,
-      tables: true,
-    },
-  });
-
-  const bookingTablesObject: BookingTablesObjectType = {};
-
-  bookings.forEach((booking) => {
-    bookingTablesObject[booking.booking_time.toISOString()] =
-      booking.tables.reduce((object, table) => {
-        return {
-          ...object,
-          [table.table_id]: true,
-        };
-      }, {});
-  });
 
   const restaurant = await prisma.restaurant.findUnique({
     where: {
@@ -73,29 +35,16 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
     return;
   }
 
-  const tables = restaurant.tables;
-
-  const searchTimesWithTables = searchTimes.map((searchTime) => {
-    return {
-      date: new Date(`${day}T${searchTime}`),
-      time: searchTime,
-      tables,
-    };
+  const searchTimesWithTables = await findAvailableTables({
+    day,
+    time,
+    restaurant,
+    res,
   });
 
-  searchTimesWithTables.forEach((searchTimeWithTables) => {
-    searchTimeWithTables.tables.filter((table) => {
-      if (bookingTablesObject[searchTimeWithTables.date.toISOString()]) {
-        if (
-          bookingTablesObject[searchTimeWithTables.date.toISOString()][table.id]
-        ) {
-          return false;
-        }
-      }
-
-      return true;
-    });
-  });
+  if (!searchTimesWithTables) {
+    return res.status(400).json({ message: "Invalid data provided" });
+  }
 
   const availabilities = searchTimesWithTables
     .map((searchTimeWithTables) => {
